refactor: introduce CountryCode enum and DEFAULT_COUNTRY in address-form-helpers; update usages and remove standalone constants file
