+++ packages/webcomponents/src/components/business-forms/business-form/business-core-info/business-core-info.tsx
@@ -2,7 +2,7 @@ import { Component, Host, h, Prop, State } from '@stencil/core';
 import { businessClassificationOptions } from '../../utils/business-form-options';
 import { FormController } from '../../../../ui-components/form/form';
 import { PHONE_MASKS, TAX_ID_MASKS } from '../../../../utils/form-input-masks';
-import { DEFAULT_COUNTRY } from '../../../../utils/constants';
+import { DEFAULT_COUNTRY, CountryCode } from '../../../../utils/address-form-helpers';
 import { CoreBusinessInfo, ICoreBusinessInfo } from '../../../../api/Business';
 import { heading2 } from '../../../../styles/parts';
 
@@ -57,7 +57,7 @@ export class BusinessCoreInfo {
   }
 
   private setLabelsAndMaskForCountry() {
-    const isCanadian = this.country === 'CAN';
+    const isCanadian = this.country === CountryCode.CAN;
     if (isCanadian) {
       this.taxIdLabel = TAX_ID_LABEL_CAN;
       this.taxIdHelpText = TAX_ID_HELP_CAN;
