+++ packages/webcomponents/src/utils/address-form-helpers.ts
@@ -1,18 +1,27 @@
 import StateOptions from './state-options';
 import CanadianProvinceOptions from './canadian-province-options';
 
+// Country codes used across forms and validation
+export enum CountryCode {
+  USA = 'USA',
+  CAN = 'CAN',
+}
+
+// Default country used when data omits a country
+export const DEFAULT_COUNTRY: CountryCode = CountryCode.USA;
+
 // Focused country options for payment provisioning (US and Canada only)
 export const PaymentProvisioningCountryOptions = [
-  { label: 'United States', value: 'USA' },
-  { label: 'Canada', value: 'CAN' },
+  { label: 'United States', value: CountryCode.USA },
+  { label: 'Canada', value: CountryCode.CAN },
 ];
 
 // Get appropriate state/province options based on selected country
 export function getRegionOptions(countryCode: string) {
   switch (countryCode) {
-    case 'USA':
+    case CountryCode.USA:
       return StateOptions;
-    case 'CAN':
+    case CountryCode.CAN:
       return CanadianProvinceOptions;
     default:
       return StateOptions; // Default to US states
@@ -22,9 +31,9 @@ export function getRegionOptions(countryCode: string) {
 // Get appropriate region label based on selected country
 export function getRegionLabel(countryCode: string) {
   switch (countryCode) {
-    case 'USA':
+    case CountryCode.USA:
       return 'State';
-    case 'CAN':
+    case CountryCode.CAN:
       return 'Province/Territory';
     default:
       return 'State';
@@ -34,9 +43,9 @@ export function getRegionLabel(countryCode: string) {
 // Get appropriate postal code label based on selected country
 export function getPostalCodeLabel(countryCode: string) {
   switch (countryCode) {
-    case 'USA':
+    case CountryCode.USA:
       return 'ZIP Code';
-    case 'CAN':
+    case CountryCode.CAN:
       return 'Postal Code';
     default:
       return 'Postal Code';
