--- packages/webcomponents/src/utils/constants.ts
@@ -1,9 +1,0 @@
-// Centralized application-wide constants
-
-/**
- * Default country code used when backend payloads omit a country.
- * Historically this was hardcoded as 'USA' in multiple places.
- */
-export const DEFAULT_COUNTRY = 'USA';
-
-
